♻️ [Refactoring]: #455 Phase 9 error カテゴリを tests/error.rs に分離
未終端 / 巻き戻し系のテスト 9 件を literal_string/tests/error.rs に
切り出し、tests.rs ハブに `mod error;` 宣言を追加。
